backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
python-docx>=1.1.0
jinja2>=3.1.0
mistune>=3.0.0
tenacity>=8.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
